Version 0.1
Most likely the final version of this project. The purpose of this project was learning the basics of python. I completed the tutorial, organized the code into separate files, and added mutiple features: sprint, character selection, level selection, user settings, multiple levels, vertical scrolling, etc.  There are still a few bugs and issues with collision, lack of try statements, and lack of proper levels/objects. However, the game runs well and the code is clean.